+++ sqlalchemy_i18n/__init__.py
@@ -10,7 +10,7 @@ __all__ = (
 )
 
 
-__version__ = '0.6.2'
+__version__ = '0.6.3'
 
 
 def make_translatable(