+++ sqlalchemy_i18n/manager.py
@@ -27,6 +27,8 @@ class TranslationManager(object):
             'base_classes': None,
             'table_name': '%s_translation',
             'locale_column_name': 'locale',
+            'default_locale': 'en',
+            'get_locale_fallback': False
         }
 
     def instrument_translatable_classes(self, mapper, cls):